feat: add permissions and delete reservation

+++ app/Http/Controllers/ReservationController.php
@@ -27,18 +27,15 @@ class ReservationController extends Controller
 
     public function create()
     {
-        // Verificar se o usuário já possui uma reserva ativa
         $hasActiveReservation = Reservation::where('user_id', Auth::id())
             ->where('reservation_time', '>=', now())
             ->exists();
 
-        // Redirecionar se o usuário já tiver uma reserva ativa
         if ($hasActiveReservation) {
             return redirect()->route('reservations.index')
                 ->with('error', 'Você já possui uma reserva ativa.');
         }
 
-        // Obter as mesas disponíveis para reserva
         $tables = Table::orderBy('number')->get();
 
         return Inertia::render('Reservations/Create', ['tables' => $tables]);
@@ -62,6 +59,12 @@ class ReservationController extends Controller
             'date' => [
                 'required',
                 'date',
+                function ($attribute, $value, $fail) {
+                    $reservationDate = Carbon::createFromFormat('Y-m-d', $value);
+                    if ($reservationDate->lt(Carbon::today())) { 
+                        $fail('Não é possível fazer reservas para datas passadas.');
+                    }
+                },
                 function ($attribute, $value, $fail) {
                     $dayOfWeek = Carbon::createFromFormat('Y-m-d', $value)->dayOfWeek;
                     if ($dayOfWeek == Carbon::SUNDAY) {
@@ -93,26 +96,33 @@ class ReservationController extends Controller
 
     public function edit(Reservation $reservation)
     {
+
+        if (Auth::id() !== $reservation->user_id) {
+            return redirect()->route('reservations.index');
+        }
+
         $tables = Table::orderBy('number')->get();
     
-        // Formatar a hora da reserva
-        $reservation->reservation_time = Carbon::createFromFormat('Y-m-d H:i:s', $reservation->reservation_time)->format('H:i');
-    
-        // Formatar a data da reserva
-        $reservation->date = Carbon::parse($reservation->reservation_time)->format('Y-m-d');
+        $reservationDateTime = Carbon::createFromFormat('Y-m-d H:i:s', $reservation->reservation_time);
+
+        $reservation->reservation_time = $reservationDateTime->format('H:i');
+
+        $reservation->date = $reservationDateTime->toDateString();
     
         return Inertia::render('Reservations/Edit', [
             'reservation' => $reservation,
             'tables' => $tables,
         ]);
     }
-    
-    
-    
 
     public function update(Request $request, Reservation $reservation)
     {
-                $validated = $request->validate([
+
+        if (Auth::id() !== $reservation->user_id) {
+            return redirect()->route('reservations.index');
+        }
+
+        $validated = $request->validate([
             'user_id' => 'required|integer',
             'table_id' => 'required|integer',
             'reservation_time' => [
@@ -128,6 +138,12 @@ class ReservationController extends Controller
             'date' => [
                 'required',
                 'date',
+                function ($attribute, $value, $fail) {
+                    $reservationDate = Carbon::createFromFormat('Y-m-d', $value);
+                    if ($reservationDate->lt(Carbon::today())) { 
+                        $fail('Não é possível fazer reservas para datas passadas.');
+                    }
+                },
                 function ($attribute, $value, $fail) {
                     $dayOfWeek = Carbon::createFromFormat('Y-m-d', $value)->dayOfWeek;
                     if ($dayOfWeek == Carbon::SUNDAY) {
@@ -154,20 +170,14 @@ class ReservationController extends Controller
 
     public function destroy(Reservation $reservation)
     {
+
+        if (Auth::id() !== $reservation->user_id) {
+            return redirect()->route('reservations.index');
+        }
+        
         $reservation->delete();
 
         return redirect()->route('reservations.index');
-    }
-
-    public function admin()
-    {
-        $this->authorize('viewAny', Reservation::class);
-
-        $reservations = Reservation::with('user')->get();
-
-        return Inertia::render('Reservations/Admin', [
-            'reservations' => $reservations
-        ]);
     }
     
 }
